+++ modules/kubernetes/eks-core/modules.tf
@@ -377,36 +377,36 @@ module "prometheus" {
   flux_enabled                           = var.fluxcd_v2_enabled
   csi_secrets_store_provider_aws_enabled = var.csi_secrets_store_provider_aws_enabled
   azad_kube_proxy_enabled                = var.azad_kube_proxy_enabled
-  starboard_enabled                      = var.starboard_enabled
+  trivy_enabled                          = var.trivy_enabled
   vpa_enabled                            = var.vpa_enabled
   node_local_dns_enabled                 = var.node_local_dns_enabled
   promtail_enabled                       = var.promtail_enabled
   node_ttl_enabled                       = var.node_ttl_enabled
 }
 
-# starboard
-module "starboard_crd" {
+# trivy
+module "trivy_crd" {
   source = "../../kubernetes/helm-crd"
 
   chart_repository = "https://aquasecurity.github.io/helm-charts/"
-  chart_name       = "starboard-operator"
-  chart_version    = "0.9.1"
+  chart_name       = "trivy-operator"
+  chart_version    = "0.11.0"
 }
 
-module "starboard" {
-  depends_on = [module.opa_gatekeeper, module.starboard_crd]
+module "trivy" {
+  depends_on = [module.opa_gatekeeper, module.trivy_crd]
 
   for_each = {
-    for s in ["starboard"] :
+    for s in ["trivy"] :
     s => s
-    if var.starboard_enabled
+    if var.trivy_enabled
   }
 
-  source = "../../kubernetes/starboard"
+  source = "../../kubernetes/trivy"
 
   cloud_provider                  = "aws"
-  starboard_role_arn              = var.starboard_config.starboard_role_arn
-  trivy_role_arn                  = var.starboard_config.trivy_role_arn
+  trivy_operator_role_arn         = var.trivy_config.trivy_operator_role_arn
+  trivy_role_arn                  = var.trivy_config.trivy_role_arn
   volume_claim_storage_class_name = "gp2"
 }