+++ CHANGELOG.md
@@ -13,7 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - [#928](https://github.com/XenitAB/terraform-modules/pull/928) Enable Node TTL by default.
-- [#930](https://github.com/XenitAB/terraform-modules/pull/930) Change from starboard to trivy-operator.
+- [#933](https://github.com/XenitAB/terraform-modules/pull/933) Change from starboard to trivy-operator.
 
 ## 2023.02.1