+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [#906](https://github.com/XenitAB/terraform-modules/pull/906) Add support for kubernetes 1.25 in Azure.
 
 ### Changed
+- [#929](https://github.com/XenitAB/terraform-modules/pull/928) Make allow_gateway_transit configurable.
 
 - [#928](https://github.com/XenitAB/terraform-modules/pull/928) Enable Node TTL by default.